- fixed: due to integer overflow caused by ids coming from the Twitter API, relevant values have been changed to the datatype "float" (closes CN-293)

+++ src/corelib/js/source/service/twitter/data/TwitterUserRecord.js
@@ -28,7 +28,7 @@ com.conjoon.service.twitter.data.TwitterUserRecord = Ext.data.Record.create([
     /**
      * @type {Number} id The id of the twitter user as managed by the Twitter service.
      */
-    {name : 'id',            type : 'int'},
+    {name : 'id',            type : 'float'},
 
     /**
      * @type {String} name The name of the user as managed by the
